# blind bump `test_benchmark_batch_insert_speed()` runtime limits